// Front/src/pages/rex/type.ts
export type ProjectRexType = {
	budget: string
	color?: number
	constraint: string
	description: string
	end_date: string
	forecast: string
	functionality: string
	id?: string
	name: string
	start_date: string
	status: boolean
	team_user: string
	technology: string
	validation: string
}

export type MyRexType = {
	rexProbleme: string
	rexReussite: string
	rexAmelioration: string
	idProjet: string
}

export type RexType = {
	answer1: string
	answer2: string
	answer3: string
}